+++ internal/infrastructure/repository/postgres/payment_find.go
@@ -8,7 +8,7 @@ import (
 	"github.com/EugeneNail/motivatr-app-payment/internal/domain"
 )
 
-func (repository *PaymentRepository) Find(ctx context.Context, id int) (*domain.Payment, error) {
+func (repository *PaymentRepository) Find(ctx context.Context, id int64) (*domain.Payment, error) {
 	payment := domain.Payment{}
 
 	row := repository.db.QueryRow(`SELECT id, date, description, category, value, user_id FROM payments WHERE id = $1`, id)
